파일 유틸 열겨형 클래스를 일반 열거형으로 변경, 메트로놈 박자 간격 타입을 GLdouble로 변경

+++ GL2D/TimerUtil.cpp
@@ -37,7 +37,7 @@ void TimerUtil::Interpolate(GLfloat Value) {
 
 
 void Metronome::SetBPM(int BPM) {
-	BeatInterval = 60.0 / (GLfloat)BPM;
+	BeatInterval = 60.0 / (GLdouble)BPM;
 }
 
 void Metronome::Update(float FT) {
